Protege os endpoints PUT e DELETE com autenticação

+++ app_gestao/app.py
@@ -1,6 +1,7 @@
 from http import HTTPStatus
 
 from fastapi import Depends, FastAPI, HTTPException  # type: ignore
+from fastapi.security import OAuth2PasswordRequestForm  # type: ignore
 from sqlalchemy import select  # type: ignore
 from sqlalchemy.exc import IntegrityError  # type: ignore
 from sqlalchemy.orm import Session  # type: ignore
@@ -9,9 +10,16 @@ from app_gestao.database import get_session
 from app_gestao.models import User
 from app_gestao.schemas import (
     Message,
+    Token,
     UserList,
     UserPublic,
     UserSchema,
+)
+from app_gestao.security import (
+    create_access_token,
+    get_current_user,
+    get_password_hash,
+    verify_password,
 )
 
 app = FastAPI()
@@ -22,6 +30,31 @@ database = []
 @app.get('/', status_code=HTTPStatus.OK, response_model=Message)
 def read_root():
     return {'message': 'Olá, mundo!'}
+
+
+@app.post('/token', response_model=Token)
+def login_for_access_token(
+    form_data: OAuth2PasswordRequestForm = Depends(),
+    session: Session = Depends(get_session),
+):
+    user = session.scalar(
+        select(User).where(User.username == form_data.username)
+    )
+    if not user:
+        raise HTTPException(
+            status_code=HTTPStatus.UNAUTHORIZED,
+            detail='Incorrect username or password',
+        )
+
+    if not verify_password(form_data.password, user.password):
+        raise HTTPException(
+            status_code=HTTPStatus.UNAUTHORIZED,
+            detail='Incorrect username or password',
+        )
+
+    access_token = create_access_token(data={'sub': user.username})
+
+    return {'acess_token': access_token, 'token_type': 'bearer'}
 
 
 @app.post('/users/', status_code=HTTPStatus.CREATED, response_model=UserPublic)
@@ -44,10 +77,12 @@ def create_user(user: UserSchema, session: Session = Depends(get_session)):
                 detail='Username already exists',
             )
 
+    hashed_password = get_password_hash(user.password)
+
     db_user = User(
         username=user.username,
         email=user.email,
-        password=user.password,
+        password=hashed_password,
     )
     session.add(db_user)
     session.commit()
@@ -68,8 +103,12 @@ def read_users(
     '/users/{user_id}', status_code=HTTPStatus.OK, response_model=UserPublic
 )
 def update_user(
-    user_id: int, user: UserSchema, session: Session = Depends(get_session)
+    user_id: int,
+    user: UserSchema,
+    session: Session = Depends(get_session),
+    current_user: User = Depends(get_current_user),
 ):
+
     db_user = session.scalar(select(User).where(User.id == user_id))
     if not db_user:
         raise HTTPException(
@@ -77,10 +116,16 @@ def update_user(
             detail='User not found',
         )
 
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=HTTPStatus.FORBIDDEN,
+            detail='Not enough permissions',
+        )
+
     try:
-        db_user.username = user.username
-        db_user.email = user.email
-        db_user.password = user.password
+        current_user.username = user.username
+        current_user.email = user.email
+        current_user.password = get_password_hash(user.password)
         session.commit()
         session.refresh(db_user)
 
@@ -95,13 +140,23 @@ def update_user(
 
 
 @app.delete('/users/{user_id}', status_code=HTTPStatus.NO_CONTENT)
-def delete_user(user_id: int, session: Session = Depends(get_session)):
+def delete_user(
+    user_id: int,
+    session: Session = Depends(get_session),
+    current_user: User = Depends(get_current_user),
+):
     db_user = session.scalar(select(User).where(User.id == user_id))
 
     if not db_user:
         raise HTTPException(
             status_code=HTTPStatus.NOT_FOUND,
             detail='User not found',
+        )
+
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=HTTPStatus.FORBIDDEN,
+            detail='Not enough permissions',
         )
 
     session.delete(db_user)
